refactor:organise each tab into a directory
