+++ src/modules/users/components/TokenActivation/TokenActivationContent/index.ts
@@ -1,2 +1,1 @@
 export { default } from './TokenActivationContent';
-export { TokensTabProps } from './TokensTab';
