+++ src/modules/users/components/TokenActivation/TokensTab/TokensTab.css
@@ -1,44 +1,6 @@
-.main {
-  min-height: 511px;
-  width: 240px;
-}
 
-.tabsListContainer {
-  height: 42px;
-}
-
-.tabsList {
-  display: flex;
-  height: 100%;
-  width: 100%;
-}
-
-.tab {
-  padding: 11px 0;
-  width: 50%;
-  border: none;
+.mainDivider {
   border-bottom: 1px solid var(--temp-grey-13);
-  background-color: inherit;
-  font-weight: var(--weight-bold);
-  text-align: center;
-  color: var(--temp-grey-blue-7);
-  outline: none;
-  cursor: pointer;
-}
-
-.tab:hover {
-  color: var(--primary);
-}
-
-.tabSelected {
-  composes: tab;
-  border-bottom-color: var(--primary);
-  color: var(--primary);
-}
-
-.tabContainer {
-  height: 100%;
-  width: 100%;
 }
 
 .totalTokensContainer {
@@ -266,7 +228,7 @@
 .claimsContainer {
   display: flex;
   justify-content: center;
-  align-items: flex-start;
+  align-items: center;
   height: 442px;
   width: 100%;
   font-size: var(--size-smallish);
@@ -279,70 +241,20 @@
   text-align: center;
 }
 
-.claimsContent {
+.stakesTabTitle {
   display: flex;
-  justify-content: center;
-  flex-direction: column;
-  padding: 15px 0;
-  width: 100%;
-}
-
-.claimAllButtonSection {
-  display: flex;
+  flex-direction: row;
   justify-content: space-between;
   align-items: center;
-  align-self: flex-end;
-  margin-bottom: 20px;
-  padding-right: 12px;
-  padding-left: 21px;
+  padding-right: 21px;
+  padding-left: 30px;
   width: 100%;
 }
 
-.claimAllButtonSection button {
-  width: 100px;
-  font-size: var(--size-small);
-}
-
-.claimAllButtonSection h4 {
-  font-size: var(--size-smallish);
-}
-
-.claimsContent span {
-  text-align: center;
-
-  .stakesTabTitle {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    padding-right: 21px;
-    padding-left: 30px;
-    width: 100%;
-  }
-
-  .dot {
-    height: 20px;
-    width: 20px;
-    border-radius: 50%;
-    background-color: var(--pink);
-    color: var(--text-invert);
-  }
-
-  .stakesTabTitle {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    padding-right: 21px;
-    padding-left: 30px;
-    width: 100%;
-  }
-
-  .dot {
-    height: 20px;
-    width: 20px;
-    border-radius: 50%;
-    background-color: var(--pink);
-    color: var(--text-invert);
-  }
+.dot {
+  height: 20px;
+  width: 20px;
+  border-radius: 50%;
+  background-color: var(--pink);
+  color: var(--text-invert);
 }
