+++ src/modules/users/components/TokenActivation/TokensTab/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TokensTab';
